+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 * Fix `RSpec/InstanceVariable` detection inside custom matchers. ([@pirj][])
+* Fix `RSpec/ScatteredSetup` to distinguish hooks with different metadata. ([@pirj][])
 
 ## 1.37.1 (2019-12-16)